♻️ refactor: 회원 정보 Member.java 로 통합
