+++ src/main/java/agaig/justeat/domain/Member.java
@@ -1,18 +1,26 @@
 package agaig.justeat.domain;
 
 public class Member {
-    private Long id;
+    private Long member_id;
     private String email;
     private String password;
     private String name;
-    private String phone;
+    private Long phone;
+    private String address;
+    private String height;
+    private String weight;
+    private String gender;
+    private String birth;
+    private String goal;
+    private String nutrient;
+    private String exercise;
 
-    public Long getId() {
-        return id;
+    public Long getMember_id() {
+        return member_id;
     }
 
-    public void setId(Long id) {
-        this.id = id;
+    public void setMember_id(Long member_id) {
+        this.member_id = member_id;
     }
 
     public String getEmail() {
@@ -39,11 +47,11 @@ public class Member {
         this.name = name;
     }
 
-    public String getPhone() {
+    public Long getPhone() {
         return phone;
     }
 
-    public void setPhone(String phone) {
+    public void setPhone(Long phone) {
         this.phone = phone;
     }
 
@@ -55,5 +63,59 @@ public class Member {
         this.address = address;
     }
 
-    private String address;
+    public String getHeight() {
+        return height;
+    }
+
+    public void setHeight(String height) {
+        this.height = height;
+    }
+
+    public String getWeight() {
+        return weight;
+    }
+
+    public void setWeight(String weight) {
+        this.weight = weight;
+    }
+
+    public String getGender() {
+        return gender;
+    }
+
+    public void setGender(String gender) {
+        this.gender = gender;
+    }
+
+    public String getBirth() {
+        return birth;
+    }
+
+    public void setBirth(String birth) {
+        this.birth = birth;
+    }
+
+    public String getGoal() {
+        return goal;
+    }
+
+    public void setGoal(String goal) {
+        this.goal = goal;
+    }
+
+    public String getNutrient() {
+        return nutrient;
+    }
+
+    public void setNutrient(String nutrient) {
+        this.nutrient = nutrient;
+    }
+
+    public String getExercise() {
+        return exercise;
+    }
+
+    public void setExercise(String exercise) {
+        this.exercise = exercise;
+    }
 }
