--- src/main/java/agaig/justeat/domain/MemberHealth.java
@@ -1,94 +1,0 @@
-package agaig.justeat.domain;
-
-public class MemberHealth {
-    private Long member_id;
-    private String height;
-    private String weight;
-    private String sex;
-    private String birth;
-    private String goal;
-    private String protein;
-    private String carbo;
-    private String fat;
-    private String exercise;
-
-    public String getExercise() {
-        return exercise;
-    }
-
-    public void setExercise(String exercise) {
-        this.exercise = exercise;
-    }
-
-    public Long getMember_id() {
-        return member_id;
-    }
-
-    public void setMember_id(Long member_id) {
-        this.member_id = member_id;
-    }
-
-    public String getHeight() {
-        return height;
-    }
-
-    public void setHeight(String height) {
-        this.height = height;
-    }
-
-    public String getWeight() {
-        return weight;
-    }
-
-    public void setWeight(String weight) {
-        this.weight = weight;
-    }
-
-    public String getSex() {
-        return sex;
-    }
-
-    public void setSex(String sex) {
-        this.sex = sex;
-    }
-
-    public String getBirth() {
-        return birth;
-    }
-
-    public void setBirth(String birth) {
-        this.birth = birth;
-    }
-
-    public String getGoal() {
-        return goal;
-    }
-
-    public void setGoal(String goal) {
-        this.goal = goal;
-    }
-
-    public String getProtein() {
-        return protein;
-    }
-
-    public void setProtein(String protein) {
-        this.protein = protein;
-    }
-
-    public String getCarbo() {
-        return carbo;
-    }
-
-    public void setCarbo(String carbo) {
-        this.carbo = carbo;
-    }
-
-    public String getFat() {
-        return fat;
-    }
-
-    public void setFat(String fat) {
-        this.fat = fat;
-    }
-}
